Rename SourcePathTool to HashedFileTool
Summary:
All code that creates a `SourcePathTool` always gives it a `PathSourcePath` and
all code that uses it just resolves the `SourcePath` to a `Path`. That means
that using a `SourcePath` is completely unnecessary and we can just replace it
with a `Path`. Because the name of the class explicitly mentions `SourcePath`,
rename it to better reflect its new functionality.

Test Plan: `buck test`