CachingBuildEngine: don't leak async jobs
Summary:
`CachingBuildEngine.build` would setup and return a `ListenableFuture`
chain to which encompassed all jobs related to the build.  However,
the future chain would return early, before pushing artifacts to the
cache, which meant we'd still actually be running stuff even after
the build had "finished".

This diff captures all the async jobs that the `CachingBuildEngine`
starts and collapses them into the uber-`Future` that gets returned.
This makes waiting on that `Future` guarantee that the build is actually
finished.

Test Plan: added unittest