C/C++: support combined preprocessing and compiling
Summary:
This diff adds support for generated combined preprocessing and compiling
rules, so that we can now do either that or separate preprocessing and
compiling (which is the default).  The logic in `CxxCompilableEnhancer`
and `CxxPreprocessoable`, which was previously responsible for generating
the separate preprocess and compile build rules, is now consolidated in
a new `CxxSourceRuleFactory` class.  This class accepts a `Strategy`
parameter so the caller can choose between separate or combined compilation.

Currently, this option is only exposed via a constructor arg to the
`CxxLibraryDescription` and `CxxBinaryDescription` classes, which was
the minimal exposure to allow all apple rules to opt-in to the combined
compilation behavior.  We should look into a better way to expose this via
the UI soon (e.g. via cxx_library rule option?).

Test Plan:
- added and updated unittest to test this behavior
- ran some compilations locally