Adds a new `audit tests` command to the CLI
Summary:
This command is meant to determine which tests should be run in various situations. For example, if someone is currently editing a file they probably want to run the tests for that files library (`buck audit tests RULE`). Once they are done they probably want to see how their project as a whole works, in which case they run the tests for the project and all of its dependencies (`buck audit dependencies --transitive RULE | xargs buck audit tests`).

Test Plan: Integration Tests.