[Easy] Remove AuditTestsOptions
Summary:
My original take on `audit tests` was to include flags for each set of tests that the user might want to see. So to see tests for all of the dependencies of a target use the `--include-dependencies` flag, and to see all the tests for rules that include the target you use `--include-owners`. @coneko pointed out that adding these flags to each every audit command wasn't a great solution, and instead suggested I do something composable. I updated the diff to remove that flag and updated the docs to suggest using `buck audit dependencies --transitive rule | xargs buck audit tests` to get the same information, but I forgot to actually remove the `AuditTestsOptions` class. This class still defined an `--include-dependencies` flag, so when I told people about this command they first tried `buck audit tests --include-dependencies RULE` and were met with sadness.

This diff fixes the issue by removing the AuditTestsOption class, so if you try using that same command you get an error instead of incorrect output.

Test Plan: Run `buck audit tests --include-dependencies RULE` before this change, see that it happily spit out the rules tests without tests from dependencies. Run the same command after this change and see an error message explaining the different available options to `audit tests` (of which `--include-dependencies` was not one of them).