[RFC] Add utility function to extract propagate `InterruptedException`
Summary:
We've gone to the trouble of decorating a lot of the functions in our
codebase to handle propagating `InterruptedException` so that our threads
are interruptible.  However, there are a few cases where a `Thread.interrupt`
causes a different type of exception.

This diff adds a `MoreThrowables.propagateInterrupt` method to wrap these
other exceptions in a `InterruptedException` and rethrows them.

Test Plan: added unittest