Create coercer for FrameworkPath.
Summary:
`AppleNativeTargetDescriptionArg.frameworks` is a set of strings, but those
strings get converted to `FrameworkPath` in some way by the code that uses
them. This conversion is being done in different places, and sometimes done
incorrectly, or supporting only a subset of all possible values. This adds
a new coercer for `FrameworkPath` that supports all source trees Buck knows
about, and changes `AppleNativeTargetDescriptionArg.frameworks` to take
advantage of it.

Test Plan: CI