Calculate the real path of a link's destination when on Windows.
Summary:
When creating links on windows, the framework likes the link's
destination to be a real path. Since this is a hard link, it doesn't
matter what type of path is used so long as it resolves to the correct
file.

Test Plan: Run `ant test -Dtest.class=AndroidAarIntegrationTest` and see it succeed.