Add windows cmd escaper code
Summary:
This is some code that we were given via email that is in the
public domain (hence the different license).  I cleaned it up
so lint passes without warnings or errors, and converted the
test to junit so we didn't have to try and run a java binary
as it was structured before.

Test Plan:
* `ant lint`
* `ant test -D"test.class"="WindowsCommandLineEscapeTest"`