New developers on Buck have no idea that 'bin' refers to a scratch directory.
Summary:
One of the first actions in code reviews for new developers on Buck
is to point out that while the rest of the world treats "bin" as an
output directory for binaries, we use the word to mean a scratch or
working directory.

Renaming the helper method in `BuildTargets` to try and resolve the
common misunderstanding. No logical changes.

Test Plan:
ant lint
buck test --all