Big Buck perf fix [2/2]: Switch AppleCxxPlatform from HashedFileTool to VersionedTool
Summary:
@andrewjcg and I discovered today that Apple builds
are extremely slow, because `AppleCxxPlatform` uses `HashedFileTool`
for `clang` and other tools.

This causes `RuleKey` generation to repeatedly hash the `clang` binary
over and over. This binary is 50 MiB or more in size, which makes
builds very slow.

This diff is part 2 of 2 of the fix.

In this diff, I move `AppleCxxPlatform` from `HashedFileTool` to
`VersionedTool`, and include in the version cache key both the version
of Xcode as well as the SDK name and version.

Test Plan:
`ant java-test -Dtest.class=AppleCxxPlatformsTest`

Ran buck build //path/to:lib in Apple repository before and after
change. Confirmed build was about 2x faster. Looked at Chrome
traces, confirmed no more big gaps in builds.