Substitute environment variables for the windows command line
Summary:
Neither cmd.exe nor powershell handles the $FOO expansion; there's a
comment to the effect of not wanting to rely on environment variables for this
to work but also a comment to the effect of needing to avoid command lines that
are too long. Hopefully on Windows we avoid a command line length limit but for
now do the substitutions to the computed command line.

Test Plan:
See in buck's test log output that the environment variables are
expanded for a genrule