Use unix separator characters for paths in .buckconfig
Summary:
When creating a `BuckConfig`, we have a path that contains `\`.
However, the `StringReader` or the `Ini` treats it as an escape
character, simply escaping the next character.  We then try to
use this path later, which fails.

We should only use the unix separator in `.buckconfig` and `BUCK`
files.

Test Plan: `ant test -D"test.class"="RepositoryFactoryTest"`