Refactor string + length hashing out of TargetGraphHashing into new class StringHashing
Summary:
I'm adding the ability to `DefaultFileHashCache` to support hashing the
contents of directories.

To do this, I want to re-use the logic already in `TargetGraphHashing` to hash
strings and recursive contents of directories.

This refactors the string hashing logic out of `TargetGraphHashing` into
a new class `StringHashing` and adds some tests.

Test Plan:
Unit tests included.
  ant java-test -Dtest.class=StringHashingTest
  buck test //test/com/facebook/buck/hashing:hashing
  buck test //test/com/facebook/buck/rules:build_rule