Ensure FakeProjectFilesystem directory iteration order is deterministic
Summary:
I need to write a test that confirms behavior depending on
iteration order of items in a directory.

Currently, `FakeProjectFilesystem` uses a `HashMap` to hold its files and
metadata. The iteration order of `HashMap` is not deterministic.

This diff changes it to use `LinkedHashMap` which uses insertion order
for iteration. That makes the behavior deterministic, and also enables
us to write tests which explicitly check for behavior when iteration
order of items in a directory changes.

I did the same thing with `HashSet` -> `LinkedHashSet`.

Test Plan:
New unit test added to confirm insertion order behavior.
  ant java-test -Dtest.class=FakeProjectFilesystemTest