Refactor path hashing out of TargetGraphHashing into new class PathHashing
Summary:
This is a follow-up to my diffs
`Refactor string + length hashing out of TargetGraphHashing into new class StringHashing`
and `Ensure FakeProjectFilesystem directory iteration order is deterministic`.

I'm adding the ability to `DefaultFileHashCache` to support hashing
the contents of directories.

To do this, I want to re-use the logic already in `TargetGraphHashing` to hash
strings and recursive contents of directories.

This refactors the path hashing logic out of `TargetGraphHashing` into
a new class `PathHashing` and adds some tests.

Test Plan:
Unit tests included.
  ant java-test -Dtest.class=PathHashingTest
  buck test //test/com/facebook/buck/hashing:hashing
  buck test //test/com/facebook/buck/rules:build_rule