Pass the appropriate environment around in `ParserTest`
Summary:
This test class was often creating `FakeBuckConfig`s, which do the
right thing with regards to the environment, but then passing
`ImmutableMap.<String, String>.of()` to methods that needed the
environment, which fails miserable on Windows.

This change cleans this up by passing the environment that the
config already knows about around.

Test Plan:
Ran the test in IntelliJ and saw no more errors about being
unable to find `python`