Make `CachingBuildEngine` own the `ListeningExecutorService`
Summary:
Previously, the `CachingBuildEngine` relied on add callbacks to
the `ListeningExecutorService` owned by the `StepRunner`.  This
diffs moves ownership into the build engine, and passes it into
the `StepRunner` via a parameter.

This makes the build engine implementation more flexible, as it
can operate directly on the executor.

Test Plan: buck test --all