Free native deflater resources when not in use
Summary:
This saves a large amount of native memory when processing lots of
uncompressed zip entries.

AppendingZipOutputStream keeps a list of all EntryAccounting objects
that have been written to it.  This is fairly wasteful of memory because
they each have their own buffer.  This diff does not address that.

The bigger problem is that they also each have a deflater, which holds
onto a *lot* native memory.  Normally, this native memory is freed when
the entry is closed, but that code is skipped for non-deflated entries.
This resulted in huge semi-permanent native memory usage when building a
java_binary from a prebuilt_jar with many uncompressed entries.  The fix
I've implemented is to eagerly close the deflater during construction if
the entry is uncompressed.

I used a small program to repro this problem while debugging.  I'm
including it in the commit in case we need it again.

Test Plan:
Ran the test program and saw that it didn't consume a crazy amount of
memory.