Support load-average limiting of build-job concurrency
Summary: This change adds an option telling Buck (and ndk-build) not to add more work to the set of running jobs when the system load average exceeds a specified threshold. This option is useful for avoiding crushing systems doing interactive work or that have little RAM. By passing the load limit down to ndk-build, we can avoid heavy C++ compilation jobs chewing through all memory on the system.