Introduce `ant ide`.
Summary:
Both Eclipse and IntelliJ seem to have some issues around running the
immutables annotation processor automatically, but Ant does not, so we
need to rely on Ant run the annotation processor to generate the appropriate
files.

Because `ant generate-immutables generate-test-immutables` is hard to
remember, this creates an alias for this so developers can just run `ant ide`.
If we have more IDE-setup-related targets in the future, they can also
become dependencies of this target.

Included a description for the target so it shows up in the `ant -p` list.

Test Plan: Run `ant ide` and `ant -p` and saw expected behavior.