Derive CxxCompilationDatabase from CxxPreprocessAndCompile rules.
Summary:
Previously, `CxxCompilationDatabase` performed a traversal over the contents
of the `BuildRuleResolver` to find the `CxxPreprocessAndCompile` rules to
use for entries in the generated compilation database, as well as for the
`HEADER_SYMLINK_TREE_FLAVOR` and `EXPORTED_HEADER_SYMLINK_TREE_FLAVOR`
dependencies.

However, running `buck build //path/to:rule#compilation-database` does not
ensure that the necessary dependencies get graph-enhanced and included in the
action graph, which means they are not guaranteed to end up in the
`BuildRuleResolver`. This fixes things so that the appropriate dependencies
are added to the action graph and added properly as deps of the `#compilation-database`.

Test Plan:
`CxxCompilationDatabaseTest` and `CxxCompilationDatabaseIntegrationTest`.

Note how the `BuildRuleResolver.addToIndex()` calls have been removed from `CxxCompilationDatabaseTest`.