Allow DexMerger to merge multiple dex files at a time
Summary:
[Forward Port of d59442ab93534e50bc5ffa06ac5dd34197ba66a9]
Previously, this class could only merge 2 dex files, which led to N^2
runtime when merging lots of dex files. This version is about 10x
faster for my test case (about 100 dex files).

Most of the changes were just replacing "dexA", "dexB", and related
objects with lists and maps. The only complicated part was the
hand-implemented 2-way merge algorithm, which I replaced with Guava's
Iterators.mergeSorted (which required implementing a proper iterator for
the section).

Note that I overwrite the uValue variable on a match because the old
code prefers writing the second value. This should not matter since
they are presumably equal, but I figured I might as well preserve the
behavior.

Test Plan:
Once all the patches have been forward-ported replace the dx from kitkat
with this version and ensure that all tests pass.