Make the `--json` flag of `audit dependencies` group results by input
Summary:
To avoid making shell scripts call buck multiple times in a single run, most of the audit commands can take multiple rules as input. This works well, but the results are not normally grouped by the input, which makes them useless if you are trying to group commands together. The `audit input` command solved this by making the `--json` flag return a JSON map, where the keys were the libraries name and the values were the libraries input. We are going to use the same strategy to solve this problem for `audit dependencies`. Doing

    buck audit dependencies --json //example:rule1 //example:rule2

will now give you a result like

    {"//example:rule1":["//example:rule1dep","//example:anotherdep"],"//example:rule2":["//example:rule2dep"]}

Test Plan: Unit/Integration Tests