Support RuleKeyAppendable types in containers in RuleKey generation
Summary:
Currently, all keys and values of maps or multimaps in RuleKey
generation have to be "single" values (e.g. `String`, `Integer`, `Enum`, or
a handful of types like `Path`, `BuildRule`, or `SourceWithFlags`).

This builds on @shs96c's recent `@RuleKeyAppendable` work to support
any type implementing `@RuleKeyAppendable` as a key or value of a map
(or an item in a container).

Test Plan: Unit tests included. `ant java-test -Dtest.class=RuleKeyTest`