Include overall success in build report.
Summary:
Previously, one had to iterate all of the results in a build report to tell
whether it succeeded or failed. (Normally, you would use the exit code to
determine this, but the client that has access to the report may not have
access to the exit code.)

This simplifies things by adding a boolean property named `"success"` to the
top level of the report.

Test Plan:
Verified docs at http://localhost:9811/command/build.html.
Updated `http://localhost:9811/command/build.html` and verified it worked.