HttpArtifactCache: embed rule key into artifact and verify on fetch
Summary:
We can't make many assumptions about cache reliability.  One potential
failure mode is that we ask for a particular artifact but get one that
actually corresponds with another key.  This diff helps catch these
errors by embedding the rule key into the artifact when we push to the
cache and verifies it on fetch.  Also, everything we embed in the artifact
is protected by the end-to-end checksum which we can make stronger (and
cryptographically secure??) if we really want to.

Test Plan:
- Added and ran http tests.
- Ran a build which wrote to an HTTP cache and verified I got cache
  hits on a second run.