Ensure we include rule keys of BuildRule fields which implement RuleKeyAppendable
Summary:
In my commit `Support RuleKeyAppendable types in containers in RuleKey generation`,
I broke the case where a `BuildRule` contains a field or dep which is a `BuildRule`
that:

1) Implements `RuleKeyAppendable`, and
2) Contains a field decorated with `@AddToRuleKey`

I mistakenly thought that any object which implemented
`RuleKeyAppendable` would express all of its rule key logic in
`appendToRuleKey()`, but that's not the case for `BuildRule`s which
use `@AddToRuleKey` on their fields.

In that case, the logic I added in my commit would cause us to invoke
`appendToRuleKey()` on the `BuildRule` without actually adding that
`BuildRule`'s rule key to the containing object.

This fixes that by ensuring we call `RuleKey.addSingleValue()` on any
`BuildRule`s which also implement `RuleKeyAppendable`.

Test Plan:
Added unit test. Confirmed it failed with `ant java-test -Dtest.class=RuleKeyTest`.

Applied fix, confirmed test didn't fail.