Fix variable expansion on Windows
Summary:
We try to expand `$VAR` and `${VAR}` on Windows for environment
variables so you can, in theory, write one command that works for
all platforms with a uniform syntax.  However, that code had a
subtle bug that was making a bunch of our tests that rely on
genrules to fail because we'd only ever expand the last variable
in the environment.

Test Plan:
Wrote a genrule with more than one environment variable in it
and saw it magically expanded properly with the debugger.