Fix Preprocessor Paths on Windows
Summary:
This fixes a fun issue with with paths on Windows in our preprocessed
files for native code.  `Paths.get("some/path")` will be represented
as `some\path` when you call `toString()` on it.  Note that the
backslash is not escaped in this string, so if you were to write it
out to a file, like we do when we postprocess our preprocessed files,
the backslash will simply escape the next character.

This change adds a method and `Function` on `Escaper` to handle this
correctly.

Test Plan:
* `ant test -D"test.class"="EscaperTest"`
* `ant test -D"test.class"="CxxPreprocessAndCompileStepTest"` (was previously failing on Windows)
* Ran `AndroidBinaryIntegrationTest.testStaticCxxLibraryDep` in IntelliJ, and see it get further