Remove `CacheMode` option
Summary:
This removes the ability to prevent a `BuildRule` from caching it's
output.  This was only ever used by `SymlinkTree` becuase we didn't
support symlinks in our artifact zips, and is replaced by performing
the actions as post build steps, which means we now report these rule
as a cache hit (instead of, effectively, a cache miss).

Test Plan: `buck test --all`