Support application-hosted tests in AppleTest
Summary:
Previously, I added support to `apple_test` to natively
build and run "logic tests" (unit tests which don't talk to the
iOS simulator).

In this diff, I add a new field `test_host_app` to `apple_test()`, which
when set to a build target pointing to an `apple_bundle()` rule, natively
builds the host application and links the test using the host application's
binary as a bundle loader.

With this setup, the test bundle is allowed to access any symbols of
the hosting application's binary.

We then pass to `XctoolRunTestsStep` the `{test bundle, host binary}` pair
so it launches the simulator, installs the host application, injects the test
bundle, and runs the tests from inside the host application.

Test Plan: New integration tests included. `ant java-test -Dtest.class=AppleTestIntegrationTest`