BuildInfoRecorder: use `ProjectFilesystem` for dir traversals
Summary:
Previously, the `BuildInfoRecorder` use a special directory traverser
class to walk the filesystem, which bypassed the passed in `ProjectFilesystem`
object making it hard to test via `FakeProjectFilesystem`.

This diff switches it to use the `ProjectFilesystem.walkRelativeFileTree`
method and adds necessary `FakeProjectFilesystem.walkRelativeFileTree`
and `FakeProjectFilesystem.createZip` methods to allow more thorough testing
of `BuildInfoRecorder`.

Test Plan: added unittests for `BuildInfoRecorder` and `FakeProjectFilesystem`.