Propagate `JavacOptions` deps to rules that use it
Summary:
`JavacOptions` has components that are potentially generated by other
build rules (e.g. Java JAR tool), which need to included as dependencies
to rules that run a build using it.  This diff adds a convenience method
to `JavacOptions` to expose the build rules dependencies required to use
it and adds it to the rules that need it.

Test Plan:
- added unittests
- `buck test --all`