AndroidBinaryGraphEnhancer: track deps from third-party JARs
Summary:
Previously, the `PrebuiltJar` rules that wrapped third-party JARs
weren't added as dependencies to the final `AndroidBinary` rules.
Normally this is fine, since the third-party JARs are available in
the sources tree.  However, when a `PrebuiltJar` is used to wrap a
JAR that is generated by another rule, we should maintain a dependency
somewhere to work properly with top-down building.

This diff models these third-party source JARs as `SourcePath` objects
and adds any dependencies extracted from them to the enhanced
dependencies.

Test Plan: added unittest