Throw `IOException` from `InitializableFromDisk`
Summary:
The methods for this interface (and `BuildOutputInitializer`)
do FS operations and so can frequently fail with `IOException`.
Previously, we handled this by wrapping this in a `RuntimeException`.
However, we already catch all excpetions in the `CachingBuildEngine`,
so this diff just adds `IOException` to the throw specs for these
methods to avoid extra boilerplate code to wrap and propagate these
exceptions.

Test Plan: `buck test --all`