Add a `--just-build` option build only a single action graph node
Summary:
A key to stress testing builds for the top-down build engine is building
each action node in isolation with an empty `buck-out` directory after
warming the cache.  This way we can verify that the action/rule can build
after only pulling in it's first-order deps.

However, this only works if the action has a corresponding node in the
target graph with the same name, as we need to use that name to select it
via command line arguments.  For action/rules that are graph enhanced and
don't have a corresponding target node by the same name, we have no way to
target them to run in isolation.

This diff adds a debugging flag to address this, `--just-build`, which allows
the user to isolate a single build target from the action graph to run.  This
provides a hook for comprehensive testing of top-down building.

Test Plan: added test