Make `DexProducedFromJavaLibrary` usuable without wrapped `JavaLibrary`
Summary:
The `DexProducedFromJavaLibrary` rule has public methods which expect
it's wrapped `JavaLibrary` to be available on disk.  Since this isn't
the case for top-down building (e.g. we could of pulled the
`DexProducedFromJavaLibrary` from cache to build it's immediate dependent,
in which case we wouldn't have the wrapped `JavaLibrary` available),
this diff breaks that requirement and has the `DexProducedFromJavaLibrary`
step store the information it needs from the wrapped `JavaLibrary` in
it's metadata.

Test Plan: added a unittest and updated others