Expose C++ binary / Apple app private headers to rules which test them
Summary:
It's very common in the Apple world to create a binary, put
it in an app bundle, then write an application-hosted test for that
app bundle which depends on the headers of the binary.

In my previous diff `Expose private library headers to rules which test them`, I implemented logic to expose private headers of
`cxx_library` rules to any test rules which test them.

This diff extends and generalizes this concept by introducing an
interface `NativeTestable` (implemented by `cxx_library`, `cxx_binary`,
and `apple_bundle`) which allows a C++ binary or Apple bundle to
expose its private headers to the test rule(s) which test it.

Test Plan:
Updated integration test in `AppleTestIntegrationTest` to
cover this case by moving binary and test into separate directories
and making headers private.

  ant java-test -Dtest.class=AppleTestIntegrationTest