[apple test] Support building xctool from source
Summary:
Currently, Buck `apple_test` rules use the open-source
`xctool` utility, which either has to exist in the repository
as a compiled binary, or called via a wrapper script which uses
Xcode to build the utility.

Building `xctool` with Xcode takes about 45 seconds and gives
no visible output to the user.

@k21 mentioned it'd make more sense to build xctool from
source, and I agree.

As an alternative, this builds on this pull request I sent in:

https://github.com/facebook/xctool/pull/518

to allow building xctool from source with Buck itself. This will
allow us to cache it in the HTTP cache, etc.

Because `xctool` is sensitive to the layout of its files, we have to
invoke a `genrule` which lays out the binary, its supporting
utilities, and supporting libraries in the correct directory
structure, then zips the result. We unzip it before running the tests.

Test Plan:
Updated `xctool` to commit 2978953f5b3dff49da0e57ad60497054f9023072
in Apple repo.

Added to `.buckconfig.local`:

  [apple]
    xctool_zip_dep = //path/to:xctool-zip

Ran `buck test //path/to:apple_test`, confirmed it built xctool and ran
test successfully.

If we like this approach, I can deprecate the support for running
`xctool` directly and remove the copy of the binary we have checked
in to the `test` directory.