CachingBuildEngine: execute action graph top-down
Summary:
Switch the `CachingBuildEngine` to use a top-down model when execution
the action graph, rather than the bottom-up approach that it used to
use which would process all rules in the action graph.  The algorithm
is roughly:
- Check rule to see if target is already up-to-date, if so, return
- Lookup rule key in cache, if found, return
- run `Build` on all first-order deps
- Check ABI based rule key (which requires deps) to see if target
  is up-to-date, if so, return.
- Run this rules' steps
- Push to cache

This approach should make builds generally more efficient, as we're
only fetching from cache or building locally things we *really* need
to complete the top-level goal.

Test Plan:
- `buck test --all`.
- Added a stress test script that randomly removes cached items and
  attempts rebuild to detect broken first-order deps.