Add extra test dependencies as runtime deps
Summary:
A common flow is to write a test that runs a binary built by another
rule, where the rule for the test lists a dependency on the rule that
builds the binary.  For this to work in top-down building, we need to
model this dependency as a runtime dependency.

This diff uses heuristic to make this work: any extra dependencies from
the target graph that don't contribute to the build are attached as runtime
dependencies via the `HasRuntimeDeps` interface.

Test Plan: added unittest verifying transitive runtime deps