Uglify paths only when needed
Summary: Right now, we apply the same "/"-adding filename transformation whether we need it or not. We only need this transformation if we need to allow for in-place modification of the debug section. If we are instead just modifying #line directives, we don't have any particular size limitations, and there's no sense in turning all our paths into cumbersome monstrosities.

Test Plan: Automated