Use @ANDROID_NDK_ROOT@ as the magic substitution for representing Android NDK in debug paths
Summary: When we rewrite debug paths to machine-independent form, we need to replace the machine-dependent path to the Android NDK with a machine-independent token. Until now, we've used  `ANDROID_TOOLS_ROOT` and `ANDROID_NDK_ROOT` for machine-independent tokens. (The former always points inside the latter, so it's strictly a shortcut.) This diff replaces both of these tokens with `@ANDROID_NDK_ROOT@`. The special marking makes it easier for external tools to notice that we have some kind of special token, and abandoning `ANDROID_TOOLS_ROOT` reduces the complexity of the code needed to reverse-map the machine-independent tokens back to normal paths.

Test Plan: Automated; readelf --debug-dump