[buck install] New class AppleSimulatorProfileParsing
Summary:
To implement `buck install` for Apple platforms, we need to
find the simulators installed on the host and choose one.

In my previous diff `New class SimctlListOutputParsing`, I
implemented part of this feature, but to complete it, we
need to know which architecture(s) and device types (iPhone,
iPad, etc.) are supported by the simulator.

This data is only present in property list files bundled with
Xcode, so we need to parse them to extract the data.

This introduces a new class `AppleSimulatorProfileParsing` to parse the
`profile.plist` files for each simulator, along with an immutable value type `AppleSimulatorProfile`
and an enum `AppleProductFamilyID`.

Test Plan: New unit test included. `ant java-test -Dtest.class=AppleSimulatorProfileParsingTest`