Fix assumption of watchman existing to get Travis passing
Summary:
Our TravisCI build started getting to the point where we run tests
again!  However, we have a test failure due to watchman not being
installed on the machines.

This change makes it an assumption failure if we ever don't have
`watchman` but try to run `buckd` in an integration test.

Test Plan: CI