Print stack traces of tests that have timed out
Summary:
If a junit (or robolectric) test has timed out, print a stack trace
of what it is doing to help the owner debug it.

Test Plan:
Modified `test.rule_timeout` in `.buckconfig.local`, and added this snippet to a test:

```    try {
      Thread.sleep((long) 50000000);
    } catch (InterruptedException e) {
      e.printStackTrace();
    }```

I then ran that test with Buck, and saw the stack trace printed out.