Allow FakeProcessExecutor to run a sequence of processes (even if they have the same params)
Summary:
Currently, the test utility `FakeProcessExecutor` takes as input a
`Map` of `ProcessExecutorParam` to `FakeProcess`.

Because a map can only contain one value per key, this means the class
can't currently model the same process being invoked multiple times
with different results.

This diff updates the utility to take either a Map, an Iterable, or a Function
to model any arbitrary sequences of processes executing in sequence.

Test Plan: Used in subsequent diff.