[buck install] Stop existing iOS simulator services before starting new one
Summary:
Previously, when starting the iOS simulator, we assumed the simulator
was not already running.

If the simulator was already running, there are three cases to worry about:

1) The booted simulator matches the one we want, and was launched with the same version
of Xcode as the one Buck is configured to use.

2) The booted simulator matches the one we want, but was not launched
with the same version of Xcode as the one Buck is configured to use.

3) The booted simulator does not match the one we want.

In case 1), we don't need to restart the simulator, but in cases 2) and 3) we need
to restart the simulator and any dependent launchd services.

This implements the above logic and adds lots of tests.

Test Plan: Unit tests included. `ant java-test -Dtest.class=AppleSimulatorRunningTest`