Make `RuleKey.Builder.build()` more general
Summary:
This diff moves the `RuleKeyPair` specific code into `AbstractBuildRule`
and makes the `RuleKey.Builder` class more focused on just the components
that get passed to it via it's methods.  This will make it a lot easier
to reused for generalized output-hash based rule keys.

Test Plan: `buck test --all`