Integration test to surface bug in incremental build of react native resources.
Summary:
I believe this can be fixed by including the hashes of relevant inputs to the
android_resource() rule in `getAbiKeyForDeps()`. The reason this cannot be included
in the rule key of android_resource() is that we need to actually execute the
`ReactNativeDeps` build rule first in order to compute this hash.

Test Plan: Test fails because `//js:app#dev,android_res` gets a matching ABI hit.