React Native (ios): package JS bundle and resources in apple bundle.
Summary:
With this diff, if an `apple_bundle` depends on an `ios_react_native_library`
rule transitively, the output `.app` directory will contain the single JS output
bundle as well as resources that are laid out next to the JS code.

The integration test demonstrates how this rule can be used, and how the outputs
are bundled in the `.app` directory.

Test Plan: Integration test.