Explicitly pass build SDK version to Apple linker with -sdk_version flag
Summary:
Apple ships SDKs with pathnames like:

  Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator.sdk

and includes a symbolic link with a version number in the name pointing to the SDK:

  Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator8.3.sdk

Previously, `AppleSdkDiscovery` filtered out any such symbolic links,
since we didn't want to include the same SDK twice.

Unfortunately, I found that `clang` actually parses the pathname to
the SDK to discover the compatibility version to encode in the binary:

https://gist.github.com/bgertzfield/672765c087e7f19e2b0e

This flag toggles on and off various runtime behaviors, and in particular
is how iOS chooses between iOS 6 and iOS 7 styling.

We confirmed in the source of Apple's `ld64` tool that this behavior
holds if you don't pass `-sdk_version` to the linker:

http://www.opensource.apple.com/source/ld64/ld64-241.9/src/ld/Options.cpp

This diff implements passing `-sdk_version` to the Apple linker.

Test Plan:
Unit tests updated. `ant java-test -Dtest.class=AppleSdkDiscoveryTest`

Set up a symbolic link `/Applications/xcode_6.3.app` -> `/Applications/Xcode.app`,
ran:

 sudo xcode-select -s /Applications/xcode_6.3.app/Contents/Developer

then built an Apple binary before and after change with the default
`iphonesimulator` platform.

Before fix, binary was tagged as SDK 6.3 (from the `6.3` in the pathname),
even though it was built with SDK version 8.3:

  % xcode-select -p
  /Applications/xcode_6.3.app/Contents/Developer
  % buck build :AppBinary
  [-] PROCESSING BUCK FILES...FINISHED 0.4s
  [-] BUILDING...FINISHED 0.4s (4/4 JOBS, 4 UPDATED, 0.0% CACHE HITS)
  % otool -l buck-out/gen/AppBinary/AppBinary | grep -A 3 LC_VERSION_MIN_IPHONEOS
        cmd LC_VERSION_MIN_IPHONEOS
    cmdsize 16
    version 8.3
        sdk 6.3

After fix, rebuilt and confirmed SDK was tagged as 8.3:

  % buck build :AppBinary
  [-] PROCESSING BUCK FILES...FINISHED 0.6s
  [-] BUILDING...FINISHED 0.8s (4/4 JOBS, 4 UPDATED, 0.0% CACHE HITS)
  % otool -l buck-out/gen/AppBinary/AppBinary | grep -A 3 LC_VERSION_MIN_IPHONEOS
        cmd LC_VERSION_MIN_IPHONEOS
    cmdsize 16
    version 8.3
        sdk 8.3