Pass SDK arguments for IOS platforms in cflags and cxxflags rather than tool arguments
Summary:
If an IOS app is being built with the compiler overridden in the [cxx] section
of the buckconfig, the arguments for the SDK, architecture, etc. will currently
be lost since they are stored in the arguments of the overridden compiler.

This passes those arguments in cflags if the compiler is overridden in the buck
config.

Test Plan: CI