Move super console cache stats calculation to rule finished event
Summary:
With the introduction of input-based rule key caching, we're now
potentially performing two fetches for each artifact during a build.
So this diff moves the cache hit/error ratio calculation to the rule
finished event handler, so that a artifact that misses on the normal
rule key fetch still reports a hit if found on a input-based rule key
fetch.

Test Plan: `buck test --all`