CachingBuildEngine: support input-based rule keys
Summary:
This diff adds support for handling input-based rule keys to the
`CachingBuildEngine`.  In particular, an input-based rule key will
be calculated for rules that support it, and will be used to avoid
rebuilding (if the key matched with the previous input-based rule
key) and to fetch the corresponding artifact from the cache.

Test Plan: Added unittests for input-based engine behavior.