Allow passing args to Apple apps in simulator
Summary:
Allows you to pass arguments to an iOS app being run in the simulator by passing them on the command line after `--`, like so:

```
buck install -r //path:target -- arg1 arg2
```

I don't love the `--` syntax but I consider it to be the best of bad options.

This is necessary for the new server-side snapshot test feature; we need the ability to pass a target output directory into the snapshot test generator application.

Test Plan: New unit test.