Fix tracking of PWD environment variable in ShellStep
Summary: Make sure the `$PWD` environment variable matches the working directory in which we execute a shell command, except if a `ShellStep` subclass overrides this tracking. If the repository root is a symbolic link and workingDirectory is `null`, `$PWD` and the working directly might be mismatched, producing unpredictable results from tools, some of which rely on `$PWD` and others of which use `getcwd`.

Test Plan: Automated; ShellStep tests modified