Add explicit metadata to cache API
Summary:
This is part of the process of detaching the build metadata from the
artifact, so that we can dedup the artifact in the backend cache.
This diff adds a metadata field to the cache API and the `CacheResult`
object, so that the `BuildInfoRecorder` can eventually use this to detach
certain build metadata from the build output for caching (and therefore
allow the backed to be more effective at deduplication).

Subsequent diffs will add support in the various cache implementations.

Test Plan: `buck test --all`